(1) fixed bug on SensorActivity for string format taking in temperature as int instead of double (2) fragment_home renamed garden nodes to sensors (3) DeviceViewModel increased timer from 3s to 6s to prevent too frequent fluctuations in online status on SensorActivity

+++ app/src/main/java/com/iotgroup2/matterapp/Pages/Home/Device/Sensor/SensorActivity.kt
@@ -397,7 +397,7 @@ class SensorActivity : AppCompatActivity() {
             lastTemp = deviceData.temperature.toDouble()
 
             if(unit) {
-                tempValueTxt.text = String.format("%.1f", deviceData.temperature / 100)
+                tempValueTxt.text = String.format("%.1f", deviceData.temperature.toDouble() / 100)
             } else {
                 tempValueTxt.text = String.format("%.1f", Utility.convertCelsiusToFahrenheit(deviceData.temperature.toDouble() / 100))
             }
